requirements.txt
gymnasium[atari,accept-rom-license]
ale-py
opencv-python
torch
torchvision
numpy
matplotlib
tqdm
pyyaml
